requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
py==1.11.0
pyparsing==2.4.4
pytest==7.4.4
requests==2.22.0
selenium==3.141.0
webdriver-manager==4.0.1
schedule==1.2.1
pytest-html==4.1.1